cockpit.js: split Deferred out to a separate file
This is another straight move (plus re-indentation).

The reason for this is that Channel uses it.  It's also another nice
chunk of code to separate from everything else.
